Add support for changing themes
Themes can now be set and remain permanently. As a test, I have inverted the colours in the 20th century 1st half theme. I have also added more JavaScript and I am preparing to allow for the exhibition to function with JSON providing data rather than a bunch of changing HTML pages.

Of course, we can prototype how one page looks with HTML, but we can do the rest with JS & JSON