Illustrate an initial draft of the 2035 theme
I have made a rough draft of the 2035 theme. It looks good so far, though I would say that it needs a bit of work. I want to make it innovative, but also like Hollow Knight. It currently is like an ultramodern rendition of an 'infection' theme, though I do want to make it really stand out as a 2035 theme.

While I was creating the theme, I was creating and modifying classes that will make it easier to make your theme work,