Implement the first complete draft of the exhibition page
All functionality now works on the exhibition page, from making text shorter/longer, to changing the language/content of the text. This took probably over 10 hours in one day to get functioning, but it finally works now and I am relieved and can rest peacefully.

Obviously, the QR code hover needs to be implemented—I will not be doing this.

Small stylistic changes have also been made to improve the responsiveness of this page.